ask-delphi-api/pyproject.toml
[project]
name = "ask-delphi-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Matthijs M.H.J. Neutelings", email = "mhj.neutelings@belastingdienst.nl" }
]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.3.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.31.0",
    "pytest>=8.0.0",
]
